Interpreter + Scoreboard Optimization

+++ src/main/scala/sl/IR/BlockReduce.scala
@@ -2,29 +2,44 @@ package sl.IR
 
 import sl.Settings
 
+private var debug = false
+
 class BlockReduce(var files: List[IRFile]){
     val map = files.map(f => f.getName() -> f).toMap
+    var globalChanged = false
 
-    def run(): List[IRFile] ={
+    def run(): (List[IRFile], Boolean) ={
         reduceEmpty()
         computeCallGraph()
 
         var changed = true
         while(changed){
+            reduceNoCall()
             reduceBlockCall()
             reduceDupplicate()
             changed = reduceEmpty()
+            if (changed){
+                globalChanged = true
+            }
         }
 
-        files.filter(f => !f.deleted)
+        val filtered = files.filter(f => !f.deleted)
+
+        (filtered, globalChanged || filtered.length != files.length)
+    }
+    def reduceNoCall()={
+        files.filter(f => f.calledBy.size == 0 && f.canBeDeleted() && !f.isJsonFile()).foreach(f => {
+            if (debug){println("delete " + f.getName() + " because it is not called by anyone")}
+            f.delete()
+        })
     }
     def computeCallGraph() ={
-        def apply(instr: IRTree)(implicit file: IRFile): Unit = {
+        def apply(instr: IRTree)(implicit parent: IRFile): Unit = {
             instr match {
                 case BlockCall(function, fullName) => {
                     map.get(fullName) match {
-                        case Some(file) => file.addCalledBy(function)
-                        case None => ()
+                        case Some(file) => file.addCalledBy(parent.getName())
+                        case None => (println("error: " + fullName + " not found"))
                     }
                 }
                 case e: IRExecute => apply(e.getStatements)
@@ -32,7 +47,7 @@ class BlockReduce(var files: List[IRFile]){
             }
         }
         files.filterNot(_.isJsonFile()).map(f => f.resetCallGraph())
-        files.filterNot(_.isJsonFile()).map(f => {
+        files.filterNot(f => f.isJsonFile() && !f.deleted).map(f => {
             f.getContents().foreach(instr => {
                 apply(instr)(f)
             })
@@ -63,21 +78,45 @@ class BlockReduce(var files: List[IRFile]){
     }
 
     def reduceBlockCall() ={
+        def applyTop(instr: IRTree)(implicit parent: IRFile): List[IRTree] = {
+            instr match {
+                case BlockCall(function, fullName) => {
+                    map.get(fullName) match {
+                        case Some(file) => {
+                            val size = file.getContents().length
+                            if (file.callByCount() == 1 && !file.hasSelfCall() && file.canBeDeleted()){
+                                file.delete()
+                                if (debug){println("delete " + file.getName() + " because it is only called once by "+ parent.getName())}
+                                file.getContents().flatMap(c => applyTop(c)(file))
+                            }
+                            else if (size == 0){
+                                List()
+                            }
+                            else{
+                                List(instr)
+                            }
+                        }
+                        case None => List(instr)
+                    }
+                }
+                case other => List(other)
+            }
+        }
         def apply(instr: IRTree)(implicit parent: IRFile): IRTree = {
             instr match {
                 case BlockCall(function, fullName) => {
                     map.get(fullName) match {
                         case Some(file) => {
                             val size = file.getContents().length
-                            if (file.callByCount() == 1 && size == 1 && !file.hasSelfCall()){
+                            if (file.callByCount() == 1 && size == 1 && !file.hasSelfCall() && file.canBeDeleted()){
                                 file.delete()
-                                //println("delete " + file.getName() + " because it is only called once by "+ parent.getName())
+                                if (debug){println("delete " + file.getName() + " because it is only called once by "+ parent.getName())}
                                 apply(file.getContents().head)(file)
                             }
                             else if (size == 1 && !file.hasSelfCall()){
                                 apply(file.getContents().head)(file)
                             }
-                            else if (size == 0){
+                            else if (size == 0 || file.deleted){
                                 EmptyIR
                             }
                             else{
@@ -97,13 +136,13 @@ class BlockReduce(var files: List[IRFile]){
         }
 
         files.filterNot(_.isJsonFile()).map(f => {
-            f.setContents(f.getContents().map(instr => {
+            f.setContents(f.getContents().flatMap(instr=> applyTop(instr)(f)).map(instr => {
                 apply(instr)(f)
             }).filterNot(_ == EmptyIR))
         })
 
         files.filter(_.getContents().length == 0).map(f=>
-            //println("delete " + f.getName() + " because it is empty")
+            if (debug){println("delete " + f.getName() + " because it is empty")}
             f.delete()
             )
     }
@@ -120,7 +159,7 @@ class BlockReduce(var files: List[IRFile]){
                                 case Some(f) if f != file => {
                                     changed = true
                                     file.delete()
-                                    //println("replace " + file.getName() + " with " + f.getName() +" in "+parent.getName()+ " because they are the same:\n"+file.getContents()+"\n===\n"+f.getContents())
+                                    if(debug){println("replace " + file.getName() + " with " + f.getName() +" in "+parent.getName()+ " because they are the same:\n"+file.getContents()+"\n===\n"+f.getContents())}
                                     BlockCall(Settings.target.getFunctionName(f.getName()), f.getName())
                                 }
                                 case _ => instr
